+++ tests/republish.rs
@@ -86,6 +86,7 @@ async fn main() {
         "republish/device",
         "127.0.0.1".parse().unwrap(),
         StandardClock::default(),
+        Settings::default(),
     )
     .unwrap();